Tutorial using all our packages for gravity processing (#1)
This is a slightly updated version of our 
[Transform 2021 tutorial](https://github.com/fatiando/transform21). It uses our new 
[curated datasets](https://github.com/fatiando/data) (downloading from Zenodo) 
and adds a few steps to the processing. Most notably, it now includes the 
conversion of "sea level" heights to ellipsoidal (geometric) heights using our geoid 
data, the topography grid is expanded to account for edge effects, density for the 
topographic correction needs to account for water and continental points below the
ellipsoid, and the regional/residual separation is done with deep equivalent sources
instead of a polynomial.